requirements.txt
Pillow
piexif
psd-tools
numpy
scipy
flask
